feat: create projects page skeleton using props on the Events components
