+++ src/routes/events/+page.svelte
@@ -1,7 +1,114 @@
 <script>
     import EventsPanel from './EventsPanel.svelte';
+    import ascenc10n from '$lib/events/ascension.webp';
+    import bootcamp10 from '$lib/events/bootcamp-10.webp';
+    import csflight from '$lib/events/csflight.webp';
+    import ex2023 from '$lib/events/EX-series-2023.webp';
+    import firstGa2023 from '$lib/events/first-GA-2324B.webp';
+    import gitWorkshop from '$lib/events/git-workshop.webp';
+    import pyrolympics from '$lib/events/pyrolympics.webp';
+    import yearEndGa2023 from '$lib/events/year-end-GA-2023.webp';
+
+    const filters = [
+        'All',
+        'EX Series',
+        'App Process',
+        'General Assembly',
+        'Partnership',
+        'Workshop',
+        'Other',
+    ];
+
+    const events = [
+        {
+            tag: 'Other',
+            state: 'Past',
+            type: 'Internal',
+            title: 'ASCENS10N',
+            dates: 'August 4 - August 6, 2023',
+            description:
+                'On its tenth year, UP CSI celebrates a decade of innovation and excellence with its members and alumni who became a part of this incredible journey.',
+            src: ascenc10n,
+            url: '/events/',
+        },
+        {
+            tag: 'EX Series',
+            state: 'Past',
+            type: 'External',
+            title: 'EX Series 2023',
+            dates: 'August 4 - August 6, 2023',
+            description:
+                'Aute incididunt officia exercitation ea aliqua officia culpa incididunt proident aliqua commodo velit.',
+            src: ex2023,
+            url: '/events/',
+        },
+        {
+            tag: 'App Process',
+            state: 'Past',
+            type: 'Internal',
+            title: 'CSFlight',
+            dates: 'August 4 - August 6, 2023',
+            description:
+                'Qui eu voluptate non nostrud. Et irure reprehenderit pariatur nulla laboris ipsum ut tempor magna officia tempor mollit Lorem reprehenderit.',
+            src: csflight,
+            url: '/events/',
+        },
+        {
+            tag: 'Partnership',
+            state: 'Past',
+            type: 'External',
+            title: 'Bootcamp 10.0',
+            dates: 'August 4 - August 6, 2023',
+            description: 'Nisi consequat consequat nisi ipsum. Ut dolore dolore officia ea ipsum.',
+            src: bootcamp10,
+            url: '/events/',
+        },
+        {
+            tag: 'Workshop',
+            state: 'Past',
+            type: 'External',
+            title: 'Git Workshop',
+            dates: 'August 4 - August 6, 2023',
+            description:
+                'Laboris ullamco do occaecat nostrud nisi irure. Ullamco nulla duis elit fugiat anim sint.',
+            src: gitWorkshop,
+            url: '/events/',
+        },
+        {
+            tag: 'General Assembly',
+            state: 'Past',
+            type: 'Internal',
+            title: 'Year-end General Assembly 2023',
+            dates: 'August 4 - August 6, 2023',
+            description: 'Proident in aliqua proident quis ex ad excepteur culpa sunt.',
+            src: yearEndGa2023,
+            url: '/events/',
+        },
+        {
+            tag: 'App Process',
+            state: 'Past',
+            type: 'Internal',
+            title: 'Pyrolympics',
+            dates: 'August 4 - August 6, 2023',
+            description:
+                'Elit irure elit adipisicing consectetur nisi laborum mollit esse eu dolore nostrud laborum laborum. Excepteur nulla velit aliquip eiusmod duis.',
+            src: pyrolympics,
+            url: '/events/',
+        },
+        {
+            tag: 'General Assembly',
+            state: 'Past',
+            type: 'Internal',
+            title: 'First GA 2324B',
+            dates: 'August 4 - August 6, 2023',
+            description:
+                'Culpa consectetur enim fugiat in labore laborum et adipisicing Lorem. Enim mollit excepteur est ea nulla in nulla dolor et.',
+            src: firstGa2023,
+            url: '/events/',
+        },
+    ];
 </script>
 
 <article class="prose max-w-full dark:prose-invert">
-    <EventsPanel />
+    <EventsPanel filters={filters} events={events} />
 </article>
